WIP: Move merging to separate RTStructMerger class
- read existing structure set files as RTStruct class object with the RTStructBuilder
- remove saving and export of resulting file, just return merged RTStruct object instead
